Remove dependency between test and publishing workflows
The idea was that the publishing workflow would only run if the testing
workflow succeeds. It looks like it currently doesn't work that way when
just a tag is pushed. After this change the publishing workflow can run
even if the tests fail, so the tests must be checked before adding a
tag. Hopefully this could be made to work in the future.